[DIDA] refactor. 코드정리 > cardId, nftId, postId, commentId -> Long  || userId -> Int 로 통일

+++ common/src/main/java/com/dida/common/actionhandler/CommentActionHandler.kt
@@ -2,5 +2,5 @@ package com.dida.common.actionhandler
 
 interface CommentActionHandler {
     fun onCommentMoreClicked(commentId: Long)
-    fun onCommentUserProfileClicked(userId: Long)
+    fun onCommentUserProfileClicked(userId: Int)
 }
