+++ data/src/main/java/com/dida/data/model/request/PostUserFollowRequest.kt
@@ -3,5 +3,5 @@ package com.dida.data.model.request
 import com.google.gson.annotations.SerializedName
 
 data class PostUserFollowRequest(
-    @SerializedName("userId") val userId: Long
+    @SerializedName("userId") val userId: Int
 )
