+++ data/src/main/java/com/dida/data/model/response/GetPostIdCommentsResponse.kt
@@ -3,9 +3,9 @@ package com.dida.data.model.response
 import com.google.gson.annotations.SerializedName
 
 data class GetPostIdCommentsResponse(
-    @SerializedName("commentId") val commentId: Int,
-    @SerializedName("postId") val postId: Int,
-    @SerializedName("userId") val userId: Long,
+    @SerializedName("commentId") val commentId: Long,
+    @SerializedName("postId") val postId: Long,
+    @SerializedName("userId") val userId: Int,
     @SerializedName("content") val content: String,
     @SerializedName("userName") val userName: String,
     @SerializedName("userImgUrl") val userImgUrl: String,
