+++ data/src/main/java/com/dida/data/model/response/GetPostPostIdResponse.kt
@@ -5,7 +5,7 @@ import com.google.gson.annotations.SerializedName
 data class GetPostPostIdResponse(
     @SerializedName("postId") val postId: Long,
     @SerializedName("cardId") val cardId: Long,
-    @SerializedName("userId") val userId: Long,
+    @SerializedName("userId") val userId: Int,
     @SerializedName("userName") val userName: String,
     @SerializedName("userImgUrl") val userImgUrl: String,
     @SerializedName("title") val title: String,
