+++ domain/src/main/java/com/dida/domain/model/nav/detailnft/DetailNFT.kt
@@ -12,5 +12,5 @@ data class DetailNFT(
     val title: String,
     val liked : Boolean,
     val type: String,
-    val userId: Long
+    val userId: Int
 )
