+++ domain/src/main/java/com/dida/domain/model/nav/post/Post.kt
@@ -3,7 +3,7 @@ package com.dida.domain.model.nav.post
 data class Post(
     val postId: Long,
     val cardId: Long,
-    val userId: Long,
+    val userId: Int,
     val userName: String,
     val userImgUrl: String,
     val title: String,
