+++ feature/home/src/main/java/com/dida/home/HomeNavigationAction.kt
@@ -3,7 +3,7 @@ package com.dida.home
 sealed class HomeNavigationAction {
     class NavigateToHotItem(val cardId: Long): HomeNavigationAction()
     class NavigateToHotSeller(val userId: Int) : HomeNavigationAction()
-    class NavigateToRecentNftItem(val nftId: Int): HomeNavigationAction()
+    class NavigateToRecentNftItem(val nftId: Long): HomeNavigationAction()
     class NavigateToSoldOut(val cardId: Long): HomeNavigationAction()
     class NavigateToCollection(val userId: Int) : HomeNavigationAction()
     object NavigateToHotSellerMore : HomeNavigationAction()
