+++ feature/hot-user/src/main/java/com/dida/hot_user/HotUserActionHandler.kt
@@ -1,6 +1,6 @@
 package com.dida.hot_user
 
 interface HotUserActionHandler {
-    fun onUserClicked(userId: Long)
-    fun onUserFollowed(userId: Long)
+    fun onUserClicked(userId: Int)
+    fun onUserFollowed(userId: Int)
 }
